requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
psycopg2-binary
alembic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pytest
requests